refactor: コンポーネントをUIごとに分ける
