+++ frontend/src/app/blog/section/TextSearchSection.tsx
