+++ frontend/src/app/page.tsx
@@ -1,37 +1,6 @@
-import Image from "next/image";
-import Header from "./header";
-import WelcomeSection from "./components/top/WelcomeSection";
-import ProfileSection from "./components/top/ProfileSection";
-import BlogListSection from "./components/top/BlogListSection";
-import ContactForm from "./components/top/ContactFormSection";
+// 最低限の経由でトップページを表示
+import Home from "./top/page";
 
-export default function Home() {
-  return (
-    <div className="min-h-screen bg-gray-50 font-sans">
-      <Header />
-      <main className="max-w-4xl mx-auto px-4 py-12">
-        <div className="bg-white rounded-lg shadow-sm p-8">
-          
-          <section className="border-b border-gray-200 mt-12 pb-12 text-center">
-            <WelcomeSection />
-          </section>
-          
-          <section className="border-b border-gray-200 mt-16 pb-16">
-            <ProfileSection />
-          </section>
-
-          <section className="border-b border-gray-200 mt-16 pb-16">
-            <BlogListSection />
-          </section>
-
-          {/* TODO: 成果物も簡単な見出しを作成しておく。詳しい内容は専用のページに遷移させる。 */}
-
-          <section className="mt-16 pb-16">
-            <ContactForm />
-          </section>
-        </div>
-
-      </main>
-    </div>
-  );
+export default function Main() {
+	return <Home />;
 }
