+++ frontend/src/app/profile/section/ProfileDetailSection.tsx
