+++ frontend/src/app/top/section/WelcomeSection.tsx
